Use a lib.rs in order to allow for integration test
main.rs now only use the new mainlib::main function.
We are now able to do `use mainlib;` in our integration tests.
