+++ Cargo.toml
@@ -4,8 +4,15 @@ version = "0.1.0"
 authors = ["Put yaself here etc"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "mainlib"
+path = "src/lib.rs"
 
+[[bin]]
+name = "webserver"
+path = "src/main.rs"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
 base64-url = "1.4.8"
 jelly = { path = "jelly" }
